Refactor email notification tests to self contained directory
