+++ app/test/e2e/email-notifications/glad-alerts-emails.spec.js
@@ -11,10 +11,10 @@ const AlertUrlService = require('services/alertUrlService');
 const AlertQueue = require('queues/alert.queue');
 const EmailHelpersService = require('services/emailHelpersService');
 
-const { getTestServer } = require('./utils/test-server');
-const { createSubscription } = require('./utils/helpers');
-const { createMockViirsAlertsQuery, createMockGeostore } = require('./utils/mock');
-const { ROLES } = require('./utils/test.constants');
+const { getTestServer } = require('../utils/test-server');
+const { createSubscription } = require('../utils/helpers');
+const { createMockAlertsQuery, createMockGeostore } = require('../utils/mock');
+const { ROLES } = require('../utils/test.constants');
 
 nock.disableNetConnect();
 nock.enableNetConnect(process.env.HOST_IP);
@@ -40,14 +40,14 @@ const assertSubscriptionStats = (jsonMessage, sub) => {
         .and.have.property('email').and.equal('info@vizzuality.com');
 };
 
-const bootstrapVIIRSAlertTest = () => {
+const bootstrapGLADAlertTest = () => {
     const beginDate = moment().subtract('1', 'w');
     const endDate = moment();
     process.on('unhandledRejection', (error) => should.fail(error));
     return { beginDate, endDate };
 };
 
-describe('VIIRS Fires alert emails', () => {
+describe('GLAD alert emails', () => {
 
     before(async () => {
         if (process.env.NODE_ENV !== 'test') {
@@ -63,23 +63,22 @@ describe('VIIRS Fires alert emails', () => {
         await Statistic.deleteMany({}).exec();
     });
 
-    it('Updating VIIRS Fires dataset triggers a new email being queued using the correct email template and providing the needed data', async () => {
+    it('Updating GLAD alerts dataset triggers a new email being queued using the correct email template and providing the needed data', async () => {
         const subscriptionOne = await new Subscription(createSubscription(
             ROLES.USER.id,
-            'viirs-active-fires',
+            'glad-alerts',
             { params: { geostore: '423e5dfb0448e692f97b590c61f45f22' } },
         )).save();
 
-        const { beginDate, endDate } = bootstrapVIIRSAlertTest();
-        createMockViirsAlertsQuery(3);
+        const { beginDate, endDate } = bootstrapGLADAlertTest();
+        createMockAlertsQuery(3);
 
         redisClient.on('message', (channel, message) => {
             const jsonMessage = JSON.parse(message);
             jsonMessage.should.have.property('template');
-
             switch (jsonMessage.template) {
 
-                case 'fires-notification-en':
+                case 'forest-change-notification-glads-en':
                     jsonMessage.should.have.property('sender').and.equal('gfw');
                     jsonMessage.should.have.property('data').and.be.a('object');
 
@@ -88,55 +87,55 @@ describe('VIIRS Fires alert emails', () => {
                         .and.have.property('address')
                         .and.have.property('email')
                         .and.equal('subscription-recipient@vizzuality.com');
-                    jsonMessage.data.should.have.property('viirs_frequency').and.equal('average');
+                    jsonMessage.data.should.have.property('glad_frequency').and.equal('average');
                     jsonMessage.data.should.have.property('month').and.equal(beginDate.format('MMMM'));
                     jsonMessage.data.should.have.property('year').and.equal(beginDate.format('YYYY'));
                     jsonMessage.data.should.have.property('week_of').and.equal(`${beginDate.format('DD MMM')}`);
                     jsonMessage.data.should.have.property('week_start').and.equal(beginDate.format('DD/MM/YYYY'));
                     jsonMessage.data.should.have.property('week_end').and.equal(endDate.format('DD/MM/YYYY'));
                     jsonMessage.data.should.have.property('priority_areas').and.deep.equal({
-                        intact_forest: 41,
-                        primary_forest: 41,
-                        peat: 1171,
-                        protected_areas: 1640,
-                        plantations: 81,
-                        other: 258,
+                        intact_forest: 6,
+                        primary_forest: 7,
+                        peat: 8,
+                        protected_areas: 9,
+                        plantations: 10,
+                        other: 11
                     });
-                    jsonMessage.data.should.have.property('viirs_count').and.equal(3232);
+                    jsonMessage.data.should.have.property('glad_count').and.equal(51);
                     jsonMessage.data.should.have.property('alerts').and.have.length(6).and.deep.equal([
-                        { alert_type: 'VIIRS', date: '10/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '11/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '12/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '13/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '14/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '15/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '10/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '11/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '12/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '13/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '14/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '15/10/2019 00:10 UTC' },
                     ]);
 
                     // Keeping this for backwards compatibility
-                    jsonMessage.data.should.have.property('alert_count').and.equal(3232);
+                    jsonMessage.data.should.have.property('alert_count').and.equal(51);
                     jsonMessage.data.should.have.property('alert_date_begin').and.equal(moment(beginDate).format('YYYY-MM-DD'));
                     jsonMessage.data.should.have.property('alert_date_end').and.equal(moment(endDate).format('YYYY-MM-DD'));
                     jsonMessage.data.should.have.property('alert_link').and.equal(AlertUrlService.generate(
                         subscriptionOne,
                         {
-                            name: 'viirs_fires_alerts',
-                            slug: 'viirs-active-fires',
+                            name: 'umd_as_it_happens',
+                            slug: 'glad-alerts',
                             subscription: true,
-                            datasetId: '1d3ccf9b-102e-4c0b-b2ea-2abcc712e194',
-                            layerId: '93e33932-3959-4201-b8c8-6ec0b32596e0'
+                            datasetId: 'bfd1d211-8106-4393-86c3-9e1ab2ee1b9b',
+                            layerId: '8e4a527d-1bcd-4a12-82b0-5a108ffec452'
                         },
                         beginDate,
                         endDate,
                     ));
                     jsonMessage.data.should.have.property('alert_name').and.equal(subscriptionOne.name);
-                    jsonMessage.data.should.have.property('layerSlug').and.equal('viirs-active-fires');
+                    jsonMessage.data.should.have.property('layerSlug').and.equal('glad-alerts');
                     jsonMessage.data.should.have.property('selected_area').and.equal('Custom Area');
                     jsonMessage.data.should.have.property('subscriptions_url').and.equal('http://staging.globalforestwatch.org/my-gfw?lang=en');
                     jsonMessage.data.should.have.property('unsubscribe_url').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/subscriptions/${subscriptionOne.id}/unsubscribe?redirect=true&lang=en`);
                     jsonMessage.data.should.have.property('downloadUrls');
-                    jsonMessage.data.downloadUrls.should.have.property('csv').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/viirs-active-fires/download/?period=${moment(beginDate).format('YYYY-MM-DD')},${moment(endDate).format('YYYY-MM-DD')}&aggregate_values=False&aggregate_by=False&geostore=423e5dfb0448e692f97b590c61f45f22&format=csv`);
-                    jsonMessage.data.downloadUrls.should.have.property('json').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/viirs-active-fires/download/?period=${moment(beginDate).format('YYYY-MM-DD')},${moment(endDate).format('YYYY-MM-DD')}&aggregate_values=False&aggregate_by=False&geostore=423e5dfb0448e692f97b590c61f45f22&format=json`);
-                    jsonMessage.data.should.have.property('value').and.equal(3232);
+                    jsonMessage.data.downloadUrls.should.have.property('csv').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/glad-alerts/download/?period=${moment(beginDate).format('YYYY-MM-DD')},${moment(endDate).format('YYYY-MM-DD')}&gladConfirmOnly=False&aggregate_values=False&aggregate_by=False&geostore=423e5dfb0448e692f97b590c61f45f22&format=csv`);
+                    jsonMessage.data.downloadUrls.should.have.property('json').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/glad-alerts/download/?period=${moment(beginDate).format('YYYY-MM-DD')},${moment(endDate).format('YYYY-MM-DD')}&gladConfirmOnly=False&aggregate_values=False&aggregate_by=False&geostore=423e5dfb0448e692f97b590c61f45f22&format=json`);
+                    jsonMessage.data.should.have.property('value').and.equal(51);
                     break;
                 case 'subscriptions-stats':
                     assertSubscriptionStats(jsonMessage, subscriptionOne);
@@ -149,23 +148,23 @@ describe('VIIRS Fires alert emails', () => {
         });
 
         await AlertQueue.processMessage(null, JSON.stringify({
-            layer_slug: 'viirs-active-fires',
+            layer_slug: 'glad-alerts',
             begin_date: beginDate,
             end_date: endDate
         }));
     });
 
-    it('Updating VIIRS Fires dataset triggers a new email being queued using the correct email template and providing the needed data taking into account language differences (FR)', async () => {
+    it('Updating GLAD alerts dataset triggers a new email being queued using the correct email template and providing the needed data taking into account language differences (FR)', async () => {
         EmailHelpersService.updateMonthTranslations();
         moment.locale('fr');
         const subscriptionOne = await new Subscription(createSubscription(
             ROLES.USER.id,
-            'viirs-active-fires',
+            'glad-alerts',
             { params: { geostore: '423e5dfb0448e692f97b590c61f45f22' }, language: 'fr' },
         )).save();
 
-        const { beginDate, endDate } = bootstrapVIIRSAlertTest();
-        createMockViirsAlertsQuery(3);
+        const { beginDate, endDate } = bootstrapGLADAlertTest();
+        createMockAlertsQuery(3);
 
         redisClient.on('message', (channel, message) => {
             const jsonMessage = JSON.parse(message);
@@ -174,7 +173,7 @@ describe('VIIRS Fires alert emails', () => {
 
             switch (jsonMessage.template) {
 
-                case 'fires-notification-fr':
+                case 'forest-change-notification-glads-fr':
                     jsonMessage.should.have.property('sender').and.equal('gfw');
                     jsonMessage.should.have.property('data').and.be.a('object');
 
@@ -183,55 +182,55 @@ describe('VIIRS Fires alert emails', () => {
                         .and.have.property('address')
                         .and.have.property('email')
                         .and.equal('subscription-recipient@vizzuality.com');
-                    jsonMessage.data.should.have.property('viirs_frequency').and.equal('moyenne');
+                    jsonMessage.data.should.have.property('glad_frequency').and.equal('moyenne');
                     jsonMessage.data.should.have.property('month').and.equal(beginDate.format('MMMM'));
                     jsonMessage.data.should.have.property('year').and.equal(beginDate.format('YYYY'));
                     jsonMessage.data.should.have.property('week_of').and.equal(`${beginDate.format('DD MMM')}`);
                     jsonMessage.data.should.have.property('week_start').and.equal(beginDate.format('DD/MM/YYYY'));
                     jsonMessage.data.should.have.property('week_end').and.equal(endDate.format('DD/MM/YYYY'));
                     jsonMessage.data.should.have.property('priority_areas').and.deep.equal({
-                        intact_forest: 41,
-                        primary_forest: 41,
-                        peat: 1171,
-                        protected_areas: 1640,
-                        plantations: 81,
-                        other: 258,
+                        intact_forest: 6,
+                        primary_forest: 7,
+                        peat: 8,
+                        protected_areas: 9,
+                        plantations: 10,
+                        other: 11
                     });
-                    jsonMessage.data.should.have.property('viirs_count').and.equal(3232);
+                    jsonMessage.data.should.have.property('glad_count').and.equal(51);
                     jsonMessage.data.should.have.property('alerts').and.have.length(6).and.deep.equal([
-                        { alert_type: 'VIIRS', date: '10/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '11/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '12/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '13/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '14/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '15/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '10/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '11/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '12/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '13/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '14/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '15/10/2019 00:10 UTC' },
                     ]);
 
                     // Keeping this for backwards compatibility
-                    jsonMessage.data.should.have.property('alert_count').and.equal(3232);
+                    jsonMessage.data.should.have.property('alert_count').and.equal(51);
                     jsonMessage.data.should.have.property('alert_date_begin').and.equal(moment(beginDate).format('YYYY-MM-DD'));
                     jsonMessage.data.should.have.property('alert_date_end').and.equal(moment(endDate).format('YYYY-MM-DD'));
                     jsonMessage.data.should.have.property('alert_link').and.equal(AlertUrlService.generate(
                         subscriptionOne,
                         {
-                            name: 'viirs_fires_alerts',
-                            slug: 'viirs-active-fires',
+                            name: 'umd_as_it_happens',
+                            slug: 'glad-alerts',
                             subscription: true,
-                            datasetId: '1d3ccf9b-102e-4c0b-b2ea-2abcc712e194',
-                            layerId: '93e33932-3959-4201-b8c8-6ec0b32596e0'
+                            datasetId: 'bfd1d211-8106-4393-86c3-9e1ab2ee1b9b',
+                            layerId: '8e4a527d-1bcd-4a12-82b0-5a108ffec452'
                         },
                         beginDate,
                         endDate,
                     ));
                     jsonMessage.data.should.have.property('alert_name').and.equal(subscriptionOne.name);
-                    jsonMessage.data.should.have.property('layerSlug').and.equal('viirs-active-fires');
+                    jsonMessage.data.should.have.property('layerSlug').and.equal('glad-alerts');
                     jsonMessage.data.should.have.property('selected_area').and.equal('Custom Area');
                     jsonMessage.data.should.have.property('subscriptions_url').and.equal(`http://staging.globalforestwatch.org/my-gfw?lang=fr`);
                     jsonMessage.data.should.have.property('unsubscribe_url').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/subscriptions/${subscriptionOne.id}/unsubscribe?redirect=true&lang=fr`);
                     jsonMessage.data.should.have.property('downloadUrls');
-                    jsonMessage.data.downloadUrls.should.have.property('csv').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/viirs-active-fires/download/?period=${moment(beginDate).format('YYYY-MM-DD')},${moment(endDate).format('YYYY-MM-DD')}&aggregate_values=False&aggregate_by=False&geostore=423e5dfb0448e692f97b590c61f45f22&format=csv`);
-                    jsonMessage.data.downloadUrls.should.have.property('json').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/viirs-active-fires/download/?period=${moment(beginDate).format('YYYY-MM-DD')},${moment(endDate).format('YYYY-MM-DD')}&aggregate_values=False&aggregate_by=False&geostore=423e5dfb0448e692f97b590c61f45f22&format=json`);
-                    jsonMessage.data.should.have.property('value').and.equal(3232);
+                    jsonMessage.data.downloadUrls.should.have.property('csv').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/glad-alerts/download/?period=${moment(beginDate).format('YYYY-MM-DD')},${moment(endDate).format('YYYY-MM-DD')}&gladConfirmOnly=False&aggregate_values=False&aggregate_by=False&geostore=423e5dfb0448e692f97b590c61f45f22&format=csv`);
+                    jsonMessage.data.downloadUrls.should.have.property('json').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/glad-alerts/download/?period=${moment(beginDate).format('YYYY-MM-DD')},${moment(endDate).format('YYYY-MM-DD')}&gladConfirmOnly=False&aggregate_values=False&aggregate_by=False&geostore=423e5dfb0448e692f97b590c61f45f22&format=json`);
+                    jsonMessage.data.should.have.property('value').and.equal(51);
                     break;
                 case 'subscriptions-stats':
                     assertSubscriptionStats(jsonMessage, subscriptionOne);
@@ -244,23 +243,23 @@ describe('VIIRS Fires alert emails', () => {
         });
 
         await AlertQueue.processMessage(null, JSON.stringify({
-            layer_slug: 'viirs-active-fires',
+            layer_slug: 'glad-alerts',
             begin_date: beginDate,
             end_date: endDate
         }));
     });
 
-    it('Updating VIIRS Fires dataset triggers a new email being queued using the correct email template and providing the needed data taking into account language differences (ZH)', async () => {
+    it('Updating GLAD alerts dataset triggers a new email being queued using the correct email template and providing the needed data taking into account language differences (ZH)', async () => {
         EmailHelpersService.updateMonthTranslations();
         moment.locale('zh');
         const subscriptionOne = await new Subscription(createSubscription(
             ROLES.USER.id,
-            'viirs-active-fires',
+            'glad-alerts',
             { params: { geostore: '423e5dfb0448e692f97b590c61f45f22' }, language: 'zh' },
         )).save();
 
-        const { beginDate, endDate } = bootstrapVIIRSAlertTest();
-        createMockViirsAlertsQuery(3);
+        const { beginDate, endDate } = bootstrapGLADAlertTest();
+        createMockAlertsQuery(3);
 
         redisClient.on('message', (channel, message) => {
             const jsonMessage = JSON.parse(message);
@@ -269,7 +268,7 @@ describe('VIIRS Fires alert emails', () => {
 
             switch (jsonMessage.template) {
 
-                case 'fires-notification-zh':
+                case 'forest-change-notification-glads-zh':
                     jsonMessage.should.have.property('sender').and.equal('gfw');
                     jsonMessage.should.have.property('data').and.be.a('object');
                     jsonMessage.should.have.property('recipients').and.be.a('array').and.length(1);
@@ -277,55 +276,55 @@ describe('VIIRS Fires alert emails', () => {
                         .and.have.property('address')
                         .and.have.property('email')
                         .and.equal('subscription-recipient@vizzuality.com');
-                    jsonMessage.data.should.have.property('viirs_frequency').and.equal('平均');
+                    jsonMessage.data.should.have.property('glad_frequency').and.equal('平均');
                     jsonMessage.data.should.have.property('month').and.equal(beginDate.format('MMMM'));
                     jsonMessage.data.should.have.property('year').and.equal(beginDate.format('YYYY'));
                     jsonMessage.data.should.have.property('week_of').and.equal(`${beginDate.format('DD MMM')}`);
                     jsonMessage.data.should.have.property('week_start').and.equal(beginDate.format('DD/MM/YYYY'));
                     jsonMessage.data.should.have.property('week_end').and.equal(endDate.format('DD/MM/YYYY'));
                     jsonMessage.data.should.have.property('priority_areas').and.deep.equal({
-                        intact_forest: 41,
-                        primary_forest: 41,
-                        peat: 1171,
-                        protected_areas: 1640,
-                        plantations: 81,
-                        other: 258,
+                        intact_forest: 6,
+                        primary_forest: 7,
+                        peat: 8,
+                        protected_areas: 9,
+                        plantations: 10,
+                        other: 11
                     });
-                    jsonMessage.data.should.have.property('viirs_count').and.equal(3232);
+                    jsonMessage.data.should.have.property('glad_count').and.equal(51);
                     jsonMessage.data.should.have.property('alerts').and.have.length(6).and.deep.equal([
-                        { alert_type: 'VIIRS', date: '10/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '11/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '12/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '13/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '14/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '15/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '10/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '11/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '12/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '13/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '14/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '15/10/2019 00:10 UTC' },
                     ]);
 
                     // Keeping this for backwards compatibility
-                    jsonMessage.data.should.have.property('alert_count').and.equal(3232);
+                    jsonMessage.data.should.have.property('alert_count').and.equal(51);
                     jsonMessage.data.should.have.property('alert_date_begin').and.equal(moment(beginDate).format('YYYY-MM-DD'));
                     jsonMessage.data.should.have.property('alert_date_end').and.equal(moment(endDate).format('YYYY-MM-DD'));
                     jsonMessage.data.should.have.property('alert_link').and.equal(AlertUrlService.generate(
                         subscriptionOne,
                         {
-                            name: 'viirs_fires_alerts',
-                            slug: 'viirs-active-fires',
+                            name: 'umd_as_it_happens',
+                            slug: 'glad-alerts',
                             subscription: true,
-                            datasetId: '1d3ccf9b-102e-4c0b-b2ea-2abcc712e194',
-                            layerId: '93e33932-3959-4201-b8c8-6ec0b32596e0'
+                            datasetId: 'bfd1d211-8106-4393-86c3-9e1ab2ee1b9b',
+                            layerId: '8e4a527d-1bcd-4a12-82b0-5a108ffec452'
                         },
                         beginDate,
                         endDate,
                     ));
                     jsonMessage.data.should.have.property('alert_name').and.equal(subscriptionOne.name);
-                    jsonMessage.data.should.have.property('layerSlug').and.equal('viirs-active-fires');
+                    jsonMessage.data.should.have.property('layerSlug').and.equal('glad-alerts');
                     jsonMessage.data.should.have.property('selected_area').and.equal('Custom Area');
                     jsonMessage.data.should.have.property('subscriptions_url').and.equal('http://staging.globalforestwatch.org/my-gfw?lang=zh');
                     jsonMessage.data.should.have.property('unsubscribe_url').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/subscriptions/${subscriptionOne.id}/unsubscribe?redirect=true&lang=zh`);
                     jsonMessage.data.should.have.property('downloadUrls');
-                    jsonMessage.data.downloadUrls.should.have.property('csv').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/viirs-active-fires/download/?period=${moment(beginDate).format('YYYY-MM-DD')},${moment(endDate).format('YYYY-MM-DD')}&aggregate_values=False&aggregate_by=False&geostore=423e5dfb0448e692f97b590c61f45f22&format=csv`);
-                    jsonMessage.data.downloadUrls.should.have.property('json').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/viirs-active-fires/download/?period=${moment(beginDate).format('YYYY-MM-DD')},${moment(endDate).format('YYYY-MM-DD')}&aggregate_values=False&aggregate_by=False&geostore=423e5dfb0448e692f97b590c61f45f22&format=json`);
-                    jsonMessage.data.should.have.property('value').and.equal(3232);
+                    jsonMessage.data.downloadUrls.should.have.property('csv').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/glad-alerts/download/?period=${moment(beginDate).format('YYYY-MM-DD')},${moment(endDate).format('YYYY-MM-DD')}&gladConfirmOnly=False&aggregate_values=False&aggregate_by=False&geostore=423e5dfb0448e692f97b590c61f45f22&format=csv`);
+                    jsonMessage.data.downloadUrls.should.have.property('json').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/glad-alerts/download/?period=${moment(beginDate).format('YYYY-MM-DD')},${moment(endDate).format('YYYY-MM-DD')}&gladConfirmOnly=False&aggregate_values=False&aggregate_by=False&geostore=423e5dfb0448e692f97b590c61f45f22&format=json`);
+                    jsonMessage.data.should.have.property('value').and.equal(51);
                     break;
                 case 'subscriptions-stats':
                     assertSubscriptionStats(jsonMessage, subscriptionOne);
@@ -338,23 +337,23 @@ describe('VIIRS Fires alert emails', () => {
         });
 
         await AlertQueue.processMessage(null, JSON.stringify({
-            layer_slug: 'viirs-active-fires',
+            layer_slug: 'glad-alerts',
             begin_date: beginDate,
             end_date: endDate
         }));
     });
 
-    it('VIIRS Fires emails for subscriptions that refer to an ISO code work as expected', async () => {
+    it('GLAD alert emails for subscriptions that refer to an ISO code work as expected', async () => {
         EmailHelpersService.updateMonthTranslations();
         moment.locale('en');
         const subscriptionOne = await new Subscription(createSubscription(
             ROLES.USER.id,
-            'viirs-active-fires',
+            'glad-alerts',
             { params: { iso: { country: 'IDN' } } },
         )).save();
 
-        const { beginDate, endDate } = bootstrapVIIRSAlertTest();
-        createMockViirsAlertsQuery(3, config.get('datasets.viirsISODataset'));
+        const { beginDate, endDate } = bootstrapGLADAlertTest();
+        createMockAlertsQuery(3, config.get('datasets.gladISODataset'));
         createMockGeostore('/v2/geostore/admin/IDN');
 
         redisClient.on('message', (channel, message) => {
@@ -364,7 +363,7 @@ describe('VIIRS Fires alert emails', () => {
 
             switch (jsonMessage.template) {
 
-                case 'fires-notification-en':
+                case 'forest-change-notification-glads-en':
                     jsonMessage.should.have.property('sender').and.equal('gfw');
                     jsonMessage.should.have.property('data').and.be.a('object');
                     jsonMessage.should.have.property('recipients').and.be.a('array').and.length(1);
@@ -372,55 +371,55 @@ describe('VIIRS Fires alert emails', () => {
                         .and.have.property('address')
                         .and.have.property('email')
                         .and.equal('subscription-recipient@vizzuality.com');
-                    jsonMessage.data.should.have.property('viirs_frequency').and.equal('average');
+                    jsonMessage.data.should.have.property('glad_frequency').and.equal('average');
                     jsonMessage.data.should.have.property('month').and.equal(beginDate.format('MMMM'));
                     jsonMessage.data.should.have.property('year').and.equal(beginDate.format('YYYY'));
                     jsonMessage.data.should.have.property('week_of').and.equal(`${beginDate.format('DD MMM')}`);
                     jsonMessage.data.should.have.property('week_start').and.equal(beginDate.format('DD/MM/YYYY'));
                     jsonMessage.data.should.have.property('week_end').and.equal(endDate.format('DD/MM/YYYY'));
                     jsonMessage.data.should.have.property('priority_areas').and.deep.equal({
-                        intact_forest: 41,
-                        primary_forest: 41,
-                        peat: 1171,
-                        protected_areas: 1640,
-                        plantations: 81,
-                        other: 258,
+                        intact_forest: 6,
+                        primary_forest: 7,
+                        peat: 8,
+                        protected_areas: 9,
+                        plantations: 10,
+                        other: 11
                     });
-                    jsonMessage.data.should.have.property('viirs_count').and.equal(3232);
+                    jsonMessage.data.should.have.property('glad_count').and.equal(51);
                     jsonMessage.data.should.have.property('alerts').and.have.length(6).and.deep.equal([
-                        { alert_type: 'VIIRS', date: '10/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '11/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '12/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '13/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '14/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '15/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '10/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '11/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '12/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '13/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '14/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '15/10/2019 00:10 UTC' },
                     ]);
 
                     // Keeping this for backwards compatibility
-                    jsonMessage.data.should.have.property('alert_count').and.equal(3232);
+                    jsonMessage.data.should.have.property('alert_count').and.equal(51);
                     jsonMessage.data.should.have.property('alert_date_begin').and.equal(moment(beginDate).format('YYYY-MM-DD'));
                     jsonMessage.data.should.have.property('alert_date_end').and.equal(moment(endDate).format('YYYY-MM-DD'));
                     jsonMessage.data.should.have.property('alert_link').and.equal(AlertUrlService.generate(
                         subscriptionOne,
                         {
-                            name: 'viirs_fires_alerts',
-                            slug: 'viirs-active-fires',
+                            name: 'umd_as_it_happens',
+                            slug: 'glad-alerts',
                             subscription: true,
-                            datasetId: '1d3ccf9b-102e-4c0b-b2ea-2abcc712e194',
-                            layerId: '93e33932-3959-4201-b8c8-6ec0b32596e0'
+                            datasetId: 'bfd1d211-8106-4393-86c3-9e1ab2ee1b9b',
+                            layerId: '8e4a527d-1bcd-4a12-82b0-5a108ffec452'
                         },
                         beginDate,
                         endDate,
                     ));
                     jsonMessage.data.should.have.property('alert_name').and.equal(subscriptionOne.name);
-                    jsonMessage.data.should.have.property('layerSlug').and.equal('viirs-active-fires');
+                    jsonMessage.data.should.have.property('layerSlug').and.equal('glad-alerts');
                     jsonMessage.data.should.have.property('selected_area').and.equal('ISO Code: IDN');
                     jsonMessage.data.should.have.property('subscriptions_url').and.equal('http://staging.globalforestwatch.org/my-gfw?lang=en');
                     jsonMessage.data.should.have.property('unsubscribe_url').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/subscriptions/${subscriptionOne.id}/unsubscribe?redirect=true&lang=en`);
                     jsonMessage.data.should.have.property('downloadUrls');
-                    jsonMessage.data.downloadUrls.should.have.property('csv').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/viirs-active-fires/download/?period=${moment(beginDate).format('YYYY-MM-DD')},${moment(endDate).format('YYYY-MM-DD')}&aggregate_values=False&aggregate_by=False&geostore=f98f505878dcee72a2e92e7510a07d6f&format=csv`);
-                    jsonMessage.data.downloadUrls.should.have.property('json').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/viirs-active-fires/download/?period=${moment(beginDate).format('YYYY-MM-DD')},${moment(endDate).format('YYYY-MM-DD')}&aggregate_values=False&aggregate_by=False&geostore=f98f505878dcee72a2e92e7510a07d6f&format=json`);
-                    jsonMessage.data.should.have.property('value').and.equal(3232);
+                    jsonMessage.data.downloadUrls.should.have.property('csv').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/glad-alerts/download/?period=${moment(beginDate).format('YYYY-MM-DD')},${moment(endDate).format('YYYY-MM-DD')}&gladConfirmOnly=False&aggregate_values=False&aggregate_by=False&geostore=f98f505878dcee72a2e92e7510a07d6f&format=csv`);
+                    jsonMessage.data.downloadUrls.should.have.property('json').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/glad-alerts/download/?period=${moment(beginDate).format('YYYY-MM-DD')},${moment(endDate).format('YYYY-MM-DD')}&gladConfirmOnly=False&aggregate_values=False&aggregate_by=False&geostore=f98f505878dcee72a2e92e7510a07d6f&format=json`);
+                    jsonMessage.data.should.have.property('value').and.equal(51);
                     break;
                 case 'subscriptions-stats':
                     assertSubscriptionStats(jsonMessage, subscriptionOne);
@@ -433,23 +432,23 @@ describe('VIIRS Fires alert emails', () => {
         });
 
         await AlertQueue.processMessage(null, JSON.stringify({
-            layer_slug: 'viirs-active-fires',
+            layer_slug: 'glad-alerts',
             begin_date: beginDate,
             end_date: endDate
         }));
     });
 
-    it('VIIRS Fires emails for subscriptions that refer to an ISO region work as expected', async () => {
+    it('GLAD alert emails for subscriptions that refer to an ISO region work as expected', async () => {
         EmailHelpersService.updateMonthTranslations();
         moment.locale('en');
         const subscriptionOne = await new Subscription(createSubscription(
             ROLES.USER.id,
-            'viirs-active-fires',
+            'glad-alerts',
             { params: { iso: { country: 'IDN', region: '3' } } },
         )).save();
 
-        const { beginDate, endDate } = bootstrapVIIRSAlertTest();
-        createMockViirsAlertsQuery(3, config.get('datasets.viirsISODataset'));
+        const { beginDate, endDate } = bootstrapGLADAlertTest();
+        createMockAlertsQuery(3, config.get('datasets.gladISODataset'));
         createMockGeostore('/v2/geostore/admin/IDN/3');
 
         redisClient.on('message', (channel, message) => {
@@ -459,7 +458,7 @@ describe('VIIRS Fires alert emails', () => {
 
             switch (jsonMessage.template) {
 
-                case 'fires-notification-en':
+                case 'forest-change-notification-glads-en':
                     jsonMessage.should.have.property('sender').and.equal('gfw');
                     jsonMessage.should.have.property('data').and.be.a('object');
                     jsonMessage.should.have.property('recipients').and.be.a('array').and.length(1);
@@ -467,55 +466,55 @@ describe('VIIRS Fires alert emails', () => {
                         .and.have.property('address')
                         .and.have.property('email')
                         .and.equal('subscription-recipient@vizzuality.com');
-                    jsonMessage.data.should.have.property('viirs_frequency').and.equal('average');
+                    jsonMessage.data.should.have.property('glad_frequency').and.equal('average');
                     jsonMessage.data.should.have.property('month').and.equal(beginDate.format('MMMM'));
                     jsonMessage.data.should.have.property('year').and.equal(beginDate.format('YYYY'));
                     jsonMessage.data.should.have.property('week_of').and.equal(`${beginDate.format('DD MMM')}`);
                     jsonMessage.data.should.have.property('week_start').and.equal(beginDate.format('DD/MM/YYYY'));
                     jsonMessage.data.should.have.property('week_end').and.equal(endDate.format('DD/MM/YYYY'));
                     jsonMessage.data.should.have.property('priority_areas').and.deep.equal({
-                        intact_forest: 41,
-                        primary_forest: 41,
-                        peat: 1171,
-                        protected_areas: 1640,
-                        plantations: 81,
-                        other: 258,
+                        intact_forest: 6,
+                        primary_forest: 7,
+                        peat: 8,
+                        protected_areas: 9,
+                        plantations: 10,
+                        other: 11
                     });
-                    jsonMessage.data.should.have.property('viirs_count').and.equal(3232);
+                    jsonMessage.data.should.have.property('glad_count').and.equal(51);
                     jsonMessage.data.should.have.property('alerts').and.have.length(6).and.deep.equal([
-                        { alert_type: 'VIIRS', date: '10/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '11/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '12/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '13/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '14/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '15/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '10/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '11/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '12/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '13/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '14/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '15/10/2019 00:10 UTC' },
                     ]);
 
                     // Keeping this for backwards compatibility
-                    jsonMessage.data.should.have.property('alert_count').and.equal(3232);
+                    jsonMessage.data.should.have.property('alert_count').and.equal(51);
                     jsonMessage.data.should.have.property('alert_date_begin').and.equal(moment(beginDate).format('YYYY-MM-DD'));
                     jsonMessage.data.should.have.property('alert_date_end').and.equal(moment(endDate).format('YYYY-MM-DD'));
                     jsonMessage.data.should.have.property('alert_link').and.equal(AlertUrlService.generate(
                         subscriptionOne,
                         {
-                            name: 'viirs_fires_alerts',
-                            slug: 'viirs-active-fires',
+                            name: 'umd_as_it_happens',
+                            slug: 'glad-alerts',
                             subscription: true,
-                            datasetId: '1d3ccf9b-102e-4c0b-b2ea-2abcc712e194',
-                            layerId: '93e33932-3959-4201-b8c8-6ec0b32596e0'
+                            datasetId: 'bfd1d211-8106-4393-86c3-9e1ab2ee1b9b',
+                            layerId: '8e4a527d-1bcd-4a12-82b0-5a108ffec452'
                         },
                         beginDate,
                         endDate,
                     ));
                     jsonMessage.data.should.have.property('alert_name').and.equal(subscriptionOne.name);
-                    jsonMessage.data.should.have.property('layerSlug').and.equal('viirs-active-fires');
+                    jsonMessage.data.should.have.property('layerSlug').and.equal('glad-alerts');
                     jsonMessage.data.should.have.property('selected_area').and.equal('ISO Code: IDN, ID1: 3');
                     jsonMessage.data.should.have.property('subscriptions_url').and.equal('http://staging.globalforestwatch.org/my-gfw?lang=en');
                     jsonMessage.data.should.have.property('unsubscribe_url').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/subscriptions/${subscriptionOne.id}/unsubscribe?redirect=true&lang=en`);
                     jsonMessage.data.should.have.property('downloadUrls');
-                    jsonMessage.data.downloadUrls.should.have.property('csv').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/viirs-active-fires/download/?period=${moment(beginDate).format('YYYY-MM-DD')},${moment(endDate).format('YYYY-MM-DD')}&aggregate_values=False&aggregate_by=False&geostore=f98f505878dcee72a2e92e7510a07d6f&format=csv`);
-                    jsonMessage.data.downloadUrls.should.have.property('json').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/viirs-active-fires/download/?period=${moment(beginDate).format('YYYY-MM-DD')},${moment(endDate).format('YYYY-MM-DD')}&aggregate_values=False&aggregate_by=False&geostore=f98f505878dcee72a2e92e7510a07d6f&format=json`);
-                    jsonMessage.data.should.have.property('value').and.equal(3232);
+                    jsonMessage.data.downloadUrls.should.have.property('csv').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/glad-alerts/download/?period=${moment(beginDate).format('YYYY-MM-DD')},${moment(endDate).format('YYYY-MM-DD')}&gladConfirmOnly=False&aggregate_values=False&aggregate_by=False&geostore=f98f505878dcee72a2e92e7510a07d6f&format=csv`);
+                    jsonMessage.data.downloadUrls.should.have.property('json').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/glad-alerts/download/?period=${moment(beginDate).format('YYYY-MM-DD')},${moment(endDate).format('YYYY-MM-DD')}&gladConfirmOnly=False&aggregate_values=False&aggregate_by=False&geostore=f98f505878dcee72a2e92e7510a07d6f&format=json`);
+                    jsonMessage.data.should.have.property('value').and.equal(51);
                     break;
                 case 'subscriptions-stats':
                     assertSubscriptionStats(jsonMessage, subscriptionOne);
@@ -528,23 +527,23 @@ describe('VIIRS Fires alert emails', () => {
         });
 
         await AlertQueue.processMessage(null, JSON.stringify({
-            layer_slug: 'viirs-active-fires',
+            layer_slug: 'glad-alerts',
             begin_date: beginDate,
             end_date: endDate
         }));
     });
 
-    it('VIIRS Fires emails for subscriptions that refer to an ISO subregion work as expected', async () => {
+    it('GLAD alert emails for subscriptions that refer to an ISO subregion work as expected', async () => {
         EmailHelpersService.updateMonthTranslations();
         moment.locale('en');
         const subscriptionOne = await new Subscription(createSubscription(
             ROLES.USER.id,
-            'viirs-active-fires',
+            'glad-alerts',
             { params: { iso: { country: 'BRA', region: '1', subregion: '1' } } },
         )).save();
 
-        const { beginDate, endDate } = bootstrapVIIRSAlertTest();
-        createMockViirsAlertsQuery(3, config.get('datasets.viirsISODataset'));
+        const { beginDate, endDate } = bootstrapGLADAlertTest();
+        createMockAlertsQuery(3, config.get('datasets.gladISODataset'));
         createMockGeostore('/v2/geostore/admin/BRA/1/1');
 
         redisClient.on('message', (channel, message) => {
@@ -554,7 +553,7 @@ describe('VIIRS Fires alert emails', () => {
 
             switch (jsonMessage.template) {
 
-                case 'fires-notification-en':
+                case 'forest-change-notification-glads-en':
                     jsonMessage.should.have.property('sender').and.equal('gfw');
                     jsonMessage.should.have.property('data').and.be.a('object');
                     jsonMessage.should.have.property('recipients').and.be.a('array').and.length(1);
@@ -562,55 +561,55 @@ describe('VIIRS Fires alert emails', () => {
                         .and.have.property('address')
                         .and.have.property('email')
                         .and.equal('subscription-recipient@vizzuality.com');
-                    jsonMessage.data.should.have.property('viirs_frequency').and.equal('average');
+                    jsonMessage.data.should.have.property('glad_frequency').and.equal('average');
                     jsonMessage.data.should.have.property('month').and.equal(beginDate.format('MMMM'));
                     jsonMessage.data.should.have.property('year').and.equal(beginDate.format('YYYY'));
                     jsonMessage.data.should.have.property('week_of').and.equal(`${beginDate.format('DD MMM')}`);
                     jsonMessage.data.should.have.property('week_start').and.equal(beginDate.format('DD/MM/YYYY'));
                     jsonMessage.data.should.have.property('week_end').and.equal(endDate.format('DD/MM/YYYY'));
                     jsonMessage.data.should.have.property('priority_areas').and.deep.equal({
-                        intact_forest: 41,
-                        primary_forest: 41,
-                        peat: 1171,
-                        protected_areas: 1640,
-                        plantations: 81,
-                        other: 258,
+                        intact_forest: 6,
+                        primary_forest: 7,
+                        peat: 8,
+                        protected_areas: 9,
+                        plantations: 10,
+                        other: 11
                     });
-                    jsonMessage.data.should.have.property('viirs_count').and.equal(3232);
+                    jsonMessage.data.should.have.property('glad_count').and.equal(51);
                     jsonMessage.data.should.have.property('alerts').and.have.length(6).and.deep.equal([
-                        { alert_type: 'VIIRS', date: '10/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '11/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '12/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '13/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '14/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '15/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '10/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '11/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '12/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '13/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '14/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '15/10/2019 00:10 UTC' },
                     ]);
 
                     // Keeping this for backwards compatibility
-                    jsonMessage.data.should.have.property('alert_count').and.equal(3232);
+                    jsonMessage.data.should.have.property('alert_count').and.equal(51);
                     jsonMessage.data.should.have.property('alert_date_begin').and.equal(moment(beginDate).format('YYYY-MM-DD'));
                     jsonMessage.data.should.have.property('alert_date_end').and.equal(moment(endDate).format('YYYY-MM-DD'));
                     jsonMessage.data.should.have.property('alert_link').and.equal(AlertUrlService.generate(
                         subscriptionOne,
                         {
-                            name: 'viirs_fires_alerts',
-                            slug: 'viirs-active-fires',
+                            name: 'umd_as_it_happens',
+                            slug: 'glad-alerts',
                             subscription: true,
-                            datasetId: '1d3ccf9b-102e-4c0b-b2ea-2abcc712e194',
-                            layerId: '93e33932-3959-4201-b8c8-6ec0b32596e0'
+                            datasetId: 'bfd1d211-8106-4393-86c3-9e1ab2ee1b9b',
+                            layerId: '8e4a527d-1bcd-4a12-82b0-5a108ffec452'
                         },
                         beginDate,
                         endDate,
                     ));
                     jsonMessage.data.should.have.property('alert_name').and.equal(subscriptionOne.name);
-                    jsonMessage.data.should.have.property('layerSlug').and.equal('viirs-active-fires');
+                    jsonMessage.data.should.have.property('layerSlug').and.equal('glad-alerts');
                     jsonMessage.data.should.have.property('selected_area').and.equal('ISO Code: BRA, ID1: 1, ID2: 1');
                     jsonMessage.data.should.have.property('subscriptions_url').and.equal('http://staging.globalforestwatch.org/my-gfw?lang=en');
                     jsonMessage.data.should.have.property('unsubscribe_url').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/subscriptions/${subscriptionOne.id}/unsubscribe?redirect=true&lang=en`);
                     jsonMessage.data.should.have.property('downloadUrls');
-                    jsonMessage.data.downloadUrls.should.have.property('csv').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/viirs-active-fires/download/?period=${moment(beginDate).format('YYYY-MM-DD')},${moment(endDate).format('YYYY-MM-DD')}&aggregate_values=False&aggregate_by=False&geostore=f98f505878dcee72a2e92e7510a07d6f&format=csv`);
-                    jsonMessage.data.downloadUrls.should.have.property('json').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/viirs-active-fires/download/?period=${moment(beginDate).format('YYYY-MM-DD')},${moment(endDate).format('YYYY-MM-DD')}&aggregate_values=False&aggregate_by=False&geostore=f98f505878dcee72a2e92e7510a07d6f&format=json`);
-                    jsonMessage.data.should.have.property('value').and.equal(3232);
+                    jsonMessage.data.downloadUrls.should.have.property('csv').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/glad-alerts/download/?period=${moment(beginDate).format('YYYY-MM-DD')},${moment(endDate).format('YYYY-MM-DD')}&gladConfirmOnly=False&aggregate_values=False&aggregate_by=False&geostore=f98f505878dcee72a2e92e7510a07d6f&format=csv`);
+                    jsonMessage.data.downloadUrls.should.have.property('json').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/glad-alerts/download/?period=${moment(beginDate).format('YYYY-MM-DD')},${moment(endDate).format('YYYY-MM-DD')}&gladConfirmOnly=False&aggregate_values=False&aggregate_by=False&geostore=f98f505878dcee72a2e92e7510a07d6f&format=json`);
+                    jsonMessage.data.should.have.property('value').and.equal(51);
                     break;
                 case 'subscriptions-stats':
                     assertSubscriptionStats(jsonMessage, subscriptionOne);
@@ -623,23 +622,23 @@ describe('VIIRS Fires alert emails', () => {
         });
 
         await AlertQueue.processMessage(null, JSON.stringify({
-            layer_slug: 'viirs-active-fires',
+            layer_slug: 'glad-alerts',
             begin_date: beginDate,
             end_date: endDate
         }));
     });
 
-    it('VIIRS Fires emails for subscriptions that refer to a WDPA ID work as expected', async () => {
+    it('GLAD alert emails for subscriptions that refer to a WDPA ID work as expected', async () => {
         EmailHelpersService.updateMonthTranslations();
         moment.locale('en');
         const subscriptionOne = await new Subscription(createSubscription(
             ROLES.USER.id,
-            'viirs-active-fires',
+            'glad-alerts',
             { params: { wdpaid: '1' } },
         )).save();
 
-        const { beginDate, endDate } = bootstrapVIIRSAlertTest();
-        createMockViirsAlertsQuery(3, config.get('datasets.viirsWDPADataset'));
+        const { beginDate, endDate } = bootstrapGLADAlertTest();
+        createMockAlertsQuery(3, config.get('datasets.gladWDPADataset'));
         createMockGeostore('/v2/geostore/wdpa/1');
 
         redisClient.on('message', (channel, message) => {
@@ -649,7 +648,7 @@ describe('VIIRS Fires alert emails', () => {
 
             switch (jsonMessage.template) {
 
-                case 'fires-notification-en':
+                case 'forest-change-notification-glads-en':
                     jsonMessage.should.have.property('sender').and.equal('gfw');
                     jsonMessage.should.have.property('data').and.be.a('object');
                     jsonMessage.should.have.property('recipients').and.be.a('array').and.length(1);
@@ -657,55 +656,55 @@ describe('VIIRS Fires alert emails', () => {
                         .and.have.property('address')
                         .and.have.property('email')
                         .and.equal('subscription-recipient@vizzuality.com');
-                    jsonMessage.data.should.have.property('viirs_frequency').and.equal('average');
+                    jsonMessage.data.should.have.property('glad_frequency').and.equal('average');
                     jsonMessage.data.should.have.property('month').and.equal(beginDate.format('MMMM'));
                     jsonMessage.data.should.have.property('year').and.equal(beginDate.format('YYYY'));
                     jsonMessage.data.should.have.property('week_of').and.equal(`${beginDate.format('DD MMM')}`);
                     jsonMessage.data.should.have.property('week_start').and.equal(beginDate.format('DD/MM/YYYY'));
                     jsonMessage.data.should.have.property('week_end').and.equal(endDate.format('DD/MM/YYYY'));
                     jsonMessage.data.should.have.property('priority_areas').and.deep.equal({
-                        intact_forest: 41,
-                        primary_forest: 41,
-                        peat: 1171,
-                        protected_areas: 1640,
-                        plantations: 81,
-                        other: 258,
+                        intact_forest: 6,
+                        primary_forest: 7,
+                        peat: 8,
+                        protected_areas: 9,
+                        plantations: 10,
+                        other: 11
                     });
-                    jsonMessage.data.should.have.property('viirs_count').and.equal(3232);
+                    jsonMessage.data.should.have.property('glad_count').and.equal(51);
                     jsonMessage.data.should.have.property('alerts').and.have.length(6).and.deep.equal([
-                        { alert_type: 'VIIRS', date: '10/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '11/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '12/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '13/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '14/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '15/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '10/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '11/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '12/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '13/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '14/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '15/10/2019 00:10 UTC' },
                     ]);
 
                     // Keeping this for backwards compatibility
-                    jsonMessage.data.should.have.property('alert_count').and.equal(3232);
+                    jsonMessage.data.should.have.property('alert_count').and.equal(51);
                     jsonMessage.data.should.have.property('alert_date_begin').and.equal(moment(beginDate).format('YYYY-MM-DD'));
                     jsonMessage.data.should.have.property('alert_date_end').and.equal(moment(endDate).format('YYYY-MM-DD'));
                     jsonMessage.data.should.have.property('alert_link').and.equal(AlertUrlService.generate(
                         subscriptionOne,
                         {
-                            name: 'viirs_fires_alerts',
-                            slug: 'viirs-active-fires',
+                            name: 'umd_as_it_happens',
+                            slug: 'glad-alerts',
                             subscription: true,
-                            datasetId: '1d3ccf9b-102e-4c0b-b2ea-2abcc712e194',
-                            layerId: '93e33932-3959-4201-b8c8-6ec0b32596e0'
+                            datasetId: 'bfd1d211-8106-4393-86c3-9e1ab2ee1b9b',
+                            layerId: '8e4a527d-1bcd-4a12-82b0-5a108ffec452'
                         },
                         beginDate,
                         endDate,
                     ));
                     jsonMessage.data.should.have.property('alert_name').and.equal(subscriptionOne.name);
-                    jsonMessage.data.should.have.property('layerSlug').and.equal('viirs-active-fires');
+                    jsonMessage.data.should.have.property('layerSlug').and.equal('glad-alerts');
                     jsonMessage.data.should.have.property('selected_area').and.equal('WDPA ID: 1');
                     jsonMessage.data.should.have.property('subscriptions_url').and.equal('http://staging.globalforestwatch.org/my-gfw?lang=en');
                     jsonMessage.data.should.have.property('unsubscribe_url').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/subscriptions/${subscriptionOne.id}/unsubscribe?redirect=true&lang=en`);
                     jsonMessage.data.should.have.property('downloadUrls');
-                    jsonMessage.data.downloadUrls.should.have.property('csv').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/viirs-active-fires/download/?period=${moment(beginDate).format('YYYY-MM-DD')},${moment(endDate).format('YYYY-MM-DD')}&aggregate_values=False&aggregate_by=False&geostore=f98f505878dcee72a2e92e7510a07d6f&format=csv`);
-                    jsonMessage.data.downloadUrls.should.have.property('json').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/viirs-active-fires/download/?period=${moment(beginDate).format('YYYY-MM-DD')},${moment(endDate).format('YYYY-MM-DD')}&aggregate_values=False&aggregate_by=False&geostore=f98f505878dcee72a2e92e7510a07d6f&format=json`);
-                    jsonMessage.data.should.have.property('value').and.equal(3232);
+                    jsonMessage.data.downloadUrls.should.have.property('csv').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/glad-alerts/download/?period=${moment(beginDate).format('YYYY-MM-DD')},${moment(endDate).format('YYYY-MM-DD')}&gladConfirmOnly=False&aggregate_values=False&aggregate_by=False&geostore=f98f505878dcee72a2e92e7510a07d6f&format=csv`);
+                    jsonMessage.data.downloadUrls.should.have.property('json').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/glad-alerts/download/?period=${moment(beginDate).format('YYYY-MM-DD')},${moment(endDate).format('YYYY-MM-DD')}&gladConfirmOnly=False&aggregate_values=False&aggregate_by=False&geostore=f98f505878dcee72a2e92e7510a07d6f&format=json`);
+                    jsonMessage.data.should.have.property('value').and.equal(51);
                     break;
                 case 'subscriptions-stats':
                     assertSubscriptionStats(jsonMessage, subscriptionOne);
@@ -718,23 +717,23 @@ describe('VIIRS Fires alert emails', () => {
         });
 
         await AlertQueue.processMessage(null, JSON.stringify({
-            layer_slug: 'viirs-active-fires',
+            layer_slug: 'glad-alerts',
             begin_date: beginDate,
             end_date: endDate
         }));
     });
 
-    it('VIIRS Fires emails for subscriptions that refer to a USE ID work as expected', async () => {
+    it('GLAD alert emails for subscriptions that refer to a USE ID work as expected', async () => {
         EmailHelpersService.updateMonthTranslations();
         moment.locale('en');
         const subscriptionOne = await new Subscription(createSubscription(
             ROLES.USER.id,
-            'viirs-active-fires',
+            'glad-alerts',
             { params: { use: 'gfw_logging', useid: '29407' } },
         )).save();
 
-        const { beginDate, endDate } = bootstrapVIIRSAlertTest();
-        createMockViirsAlertsQuery(3);
+        const { beginDate, endDate } = bootstrapGLADAlertTest();
+        createMockAlertsQuery(3);
         createMockGeostore('/v2/geostore/use/gfw_logging/29407', 4);
 
         redisClient.on('message', (channel, message) => {
@@ -744,7 +743,7 @@ describe('VIIRS Fires alert emails', () => {
 
             switch (jsonMessage.template) {
 
-                case 'fires-notification-en':
+                case 'forest-change-notification-glads-en':
                     jsonMessage.should.have.property('sender').and.equal('gfw');
                     jsonMessage.should.have.property('data').and.be.a('object');
                     jsonMessage.should.have.property('recipients').and.be.a('array').and.length(1);
@@ -752,55 +751,55 @@ describe('VIIRS Fires alert emails', () => {
                         .and.have.property('address')
                         .and.have.property('email')
                         .and.equal('subscription-recipient@vizzuality.com');
-                    jsonMessage.data.should.have.property('viirs_frequency').and.equal('average');
+                    jsonMessage.data.should.have.property('glad_frequency').and.equal('average');
                     jsonMessage.data.should.have.property('month').and.equal(beginDate.format('MMMM'));
                     jsonMessage.data.should.have.property('year').and.equal(beginDate.format('YYYY'));
                     jsonMessage.data.should.have.property('week_of').and.equal(`${beginDate.format('DD MMM')}`);
                     jsonMessage.data.should.have.property('week_start').and.equal(beginDate.format('DD/MM/YYYY'));
                     jsonMessage.data.should.have.property('week_end').and.equal(endDate.format('DD/MM/YYYY'));
                     jsonMessage.data.should.have.property('priority_areas').and.deep.equal({
-                        intact_forest: 41,
-                        primary_forest: 41,
-                        peat: 1171,
-                        protected_areas: 1640,
-                        plantations: 81,
-                        other: 258,
+                        intact_forest: 6,
+                        primary_forest: 7,
+                        peat: 8,
+                        protected_areas: 9,
+                        plantations: 10,
+                        other: 11
                     });
-                    jsonMessage.data.should.have.property('viirs_count').and.equal(3232);
+                    jsonMessage.data.should.have.property('glad_count').and.equal(51);
                     jsonMessage.data.should.have.property('alerts').and.have.length(6).and.deep.equal([
-                        { alert_type: 'VIIRS', date: '10/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '11/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '12/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '13/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '14/10/2019 00:10 UTC' },
-                        { alert_type: 'VIIRS', date: '15/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '10/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '11/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '12/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '13/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '14/10/2019 00:10 UTC' },
+                        { alert_type: 'GLAD', date: '15/10/2019 00:10 UTC' },
                     ]);
 
                     // Keeping this for backwards compatibility
-                    jsonMessage.data.should.have.property('alert_count').and.equal(3232);
+                    jsonMessage.data.should.have.property('alert_count').and.equal(51);
                     jsonMessage.data.should.have.property('alert_date_begin').and.equal(moment(beginDate).format('YYYY-MM-DD'));
                     jsonMessage.data.should.have.property('alert_date_end').and.equal(moment(endDate).format('YYYY-MM-DD'));
                     jsonMessage.data.should.have.property('alert_link').and.equal(AlertUrlService.generate(
                         subscriptionOne,
                         {
-                            name: 'viirs_fires_alerts',
-                            slug: 'viirs-active-fires',
+                            name: 'umd_as_it_happens',
+                            slug: 'glad-alerts',
                             subscription: true,
-                            datasetId: '1d3ccf9b-102e-4c0b-b2ea-2abcc712e194',
-                            layerId: '93e33932-3959-4201-b8c8-6ec0b32596e0'
+                            datasetId: 'bfd1d211-8106-4393-86c3-9e1ab2ee1b9b',
+                            layerId: '8e4a527d-1bcd-4a12-82b0-5a108ffec452'
                         },
                         beginDate,
                         endDate,
                     ));
                     jsonMessage.data.should.have.property('alert_name').and.equal(subscriptionOne.name);
-                    jsonMessage.data.should.have.property('layerSlug').and.equal('viirs-active-fires');
+                    jsonMessage.data.should.have.property('layerSlug').and.equal('glad-alerts');
                     jsonMessage.data.should.have.property('selected_area').and.equal('Custom Area');
                     jsonMessage.data.should.have.property('subscriptions_url').and.equal('http://staging.globalforestwatch.org/my-gfw?lang=en');
                     jsonMessage.data.should.have.property('unsubscribe_url').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/subscriptions/${subscriptionOne.id}/unsubscribe?redirect=true&lang=en`);
                     jsonMessage.data.should.have.property('downloadUrls');
-                    jsonMessage.data.downloadUrls.should.have.property('csv').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/viirs-active-fires/download/?period=${moment(beginDate).format('YYYY-MM-DD')},${moment(endDate).format('YYYY-MM-DD')}&aggregate_values=False&aggregate_by=False&geostore=f98f505878dcee72a2e92e7510a07d6f&format=csv`);
-                    jsonMessage.data.downloadUrls.should.have.property('json').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/viirs-active-fires/download/?period=${moment(beginDate).format('YYYY-MM-DD')},${moment(endDate).format('YYYY-MM-DD')}&aggregate_values=False&aggregate_by=False&geostore=f98f505878dcee72a2e92e7510a07d6f&format=json`);
-                    jsonMessage.data.should.have.property('value').and.equal(3232);
+                    jsonMessage.data.downloadUrls.should.have.property('csv').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/glad-alerts/download/?period=${moment(beginDate).format('YYYY-MM-DD')},${moment(endDate).format('YYYY-MM-DD')}&gladConfirmOnly=False&aggregate_values=False&aggregate_by=False&geostore=f98f505878dcee72a2e92e7510a07d6f&format=csv`);
+                    jsonMessage.data.downloadUrls.should.have.property('json').and.equal(`${process.env.API_GATEWAY_EXTERNAL_URL}/glad-alerts/download/?period=${moment(beginDate).format('YYYY-MM-DD')},${moment(endDate).format('YYYY-MM-DD')}&gladConfirmOnly=False&aggregate_values=False&aggregate_by=False&geostore=f98f505878dcee72a2e92e7510a07d6f&format=json`);
+                    jsonMessage.data.should.have.property('value').and.equal(51);
                     break;
                 case 'subscriptions-stats':
                     assertSubscriptionStats(jsonMessage, subscriptionOne);
@@ -813,21 +812,21 @@ describe('VIIRS Fires alert emails', () => {
         });
 
         await AlertQueue.processMessage(null, JSON.stringify({
-            layer_slug: 'viirs-active-fires',
+            layer_slug: 'glad-alerts',
             begin_date: beginDate,
             end_date: endDate
         }));
     });
 
-    it('No email is sent if there no alerts are returned by the VIIRS Fires alerts query', async () => {
+    it('No email is sent if there no alerts are returned by the glad alerts query', async () => {
         await new Subscription(createSubscription(
             ROLES.USER.id,
-            'viirs-active-fires',
+            'glad-alerts',
             { params: { geostore: '423e5dfb0448e692f97b590c61f45f22' } },
         )).save();
 
-        const { beginDate, endDate } = bootstrapVIIRSAlertTest();
-        createMockViirsAlertsQuery(1, undefined, { data: [] });
+        const { beginDate, endDate } = bootstrapGLADAlertTest();
+        createMockAlertsQuery(1, undefined, { data: [] });
 
         redisClient.on('message', (channel, message) => {
             const jsonMessage = JSON.parse(message);
@@ -838,7 +837,7 @@ describe('VIIRS Fires alert emails', () => {
                     jsonMessage.should.have.property('sender').and.equal('gfw');
                     jsonMessage.should.have.property('data').and.be.a('object');
                     jsonMessage.data.should.have.property('counter').and.equal(0);
-                    jsonMessage.data.should.have.property('dataset').and.equal('viirs-active-fires');
+                    jsonMessage.data.should.have.property('dataset').and.equal('glad-alerts');
                     jsonMessage.data.should.have.property('users').and.be.an('array').and.length(0);
                     jsonMessage.should.have.property('recipients').and.be.a('array').and.length(1);
                     jsonMessage.recipients[0].should.be.an('object').and.have.property('address').and.have.property('email').and.equal('info@vizzuality.com');
@@ -851,7 +850,7 @@ describe('VIIRS Fires alert emails', () => {
         });
 
         await AlertQueue.processMessage(null, JSON.stringify({
-            layer_slug: 'viirs-active-fires',
+            layer_slug: 'glad-alerts',
             begin_date: beginDate,
             end_date: endDate
         }));
